Add plugin to list user's branches
git-my requires an updated version of bash, so add it to the Brewfile.
git-my also uses the terminal colors to display a table with alternating
row highlighting, but the lack of contrast made the text hard to read.
Update one of the terminal colors used for row highlighting to improve
readability.